split telemetry rasp, waf and user files
